backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
apscheduler
httpx
python-jose[cryptography]
passlib[bcrypt]
razorpay
google-genai
pydub
numpy
lameenc
pybase64
pytest
pandas
openpyxl
deepgram-sdk>=3.0.0
websockets>=12.0
pyaudio
